refactor: modularize SourcesModal components into a dedicated file to improve maintainability
